added new compact decompositions
