+++ strawberryfields/ops.py
@@ -2357,6 +2357,9 @@ class Interferometer(Decomposition):
       To instead decompose the interferometer using the :class:`~.ops.MZgate`,
       use ``mesh='rectangular_symmetric'``.
 
+      To use the compact rectangular decomposition of Bell and Walmsley (arXiv:2104.0756), use 
+      ``mesh='rectangular_compact'``.
+
     * ``mesh='triangular'``: uses the scheme described in :cite:`reck1994`,
       resulting in a *triangular* array of :math:`M(M-1)/2` beamsplitters:
 
@@ -2364,6 +2367,9 @@ class Interferometer(Decomposition):
           :align: center
           :width: 30%
           :target: javascript:void(0);
+
+      To use the compact triangular decomposition, use
+      ``mesh='triangular_compact'``.
 
       Local phase shifts appear at the end of the beamsplitter array.
 
@@ -2382,7 +2388,13 @@ class Interferometer(Decomposition):
               placed after all interferometers, and all beamsplitters decomposed into
               pairs of symmetric beamsplitters and phase shifters
 
+            - ``rectangular_compact'`` - rectangular mesh, with two independant phase shifts
+              placed inside each MZI, extra phase shifts on edges and at the input and output.
+
             - ``'triangular'`` - triangular mesh
+
+            - ``'triangular_compact'`` - triangular mesh, with two independant phase shifts
+             placed inside each MZI.
 
         drop_identity (bool): If ``True``, decomposed gates with trivial parameters,
             such that they correspond to an identity operation, are removed.
@@ -2445,7 +2457,45 @@ class Interferometer(Decomposition):
 
         cmds = []
 
-        if not self.identity or not drop_identity:
+        if mesh == 'rectangular_compact':
+            phases = dec.rectangular_compact(self.p[0], rtol=tol, atol=tol)
+            m = phases['m']
+            for j in range(0,m-1,2):
+                phi = phases['phi_ins'][j]
+                cmds.append(Command(Rgate(phi),reg[j]))
+            for layer in range(m):
+                if (layer + m + 1) % 2 == 0:
+                    phi_bottom = phases['phi_edges'][m-1, layer]
+                    cmds.append(Command(Rgate(phi_bottom), reg[m-1]))
+                for mode in range(layer % 2, m-1, 2):
+                    delta = phases['deltas'][mode, layer]
+                    sigma = phases['sigmas'][mode, layer]
+                    phi1 = sigma+delta
+                    phi2 = sigma-delta
+                    cmds.append(Command(sMZgate(phi1, phi2), (reg[mode], reg[mode+1])))
+            for j, phi_j in phases['phi_outs'].items():
+                Rgate(phi_j) | q[j]
+                cmds.append(Command(Rgate(phi_j), reg[j]))
+
+        elif mesh == 'triangular_compact':
+            phases = dec.triangular_compact(self.p[0], rtol=tol, atol=tol)
+            m = phases['m']
+            for j in range(m-1):
+                phi_j = phases['phi_ins'][j]
+                cmds.append(Command(Rgate(phi_j),reg[j+1]))
+                for k in range(j+1):
+                    n = j - k
+                    delta = phases['deltas'][n,k]
+                    sigma = phases['sigmas'][n,k]
+                    phi1 = sigma+delta
+                    phi2 = sigma-delta
+                    cmds.append(Command(sMZgate(phi1, phi2),(reg[n], reg[n+1])))
+
+            for j in range(m):
+                zeta = phases['zetas'][j]
+                cmds.append(Command(Rgate(zeta), reg[j]))
+
+        elif not self.identity or not drop_identity:
             decomp_fn = getattr(dec, mesh)
             BS1, R, BS2 = decomp_fn(self.p[0], tol=tol)
